backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
prometheus-client==0.19.0
pymongo==4.5.0
opensearch-py==2.4.1
psycopg2-binary==2.9.9